storage weights integrated
