+++ pallets/utxo-nft/src/weights.rs
@@ -66,3 +66,9 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 			.saturating_add(Weight::from_parts(0, 9436).saturating_mul(i.into()))
 	}
 }
+
+impl WeightInfo for () {
+	fn run_process(_: u32, _: u32) -> Weight {
+			Weight::from_parts(0, 0)
+	}
+}
